add PerPageChooser component
