+++ src/modules/common/components/pagination/Pagination.js
@@ -3,14 +3,9 @@ import { withRouter } from 'react-router';
 import PropTypes from 'prop-types';
 import _ from 'underscore';
 import { router } from 'modules/common/utils';
-import { Icon, DropdownToggle } from 'modules/common/components';
-import {
-  PaginationWrapper,
-  PaginationList,
-  PerPageButton,
-  Option
-} from './styles';
-import { Dropdown } from 'react-bootstrap';
+import { Icon } from 'modules/common/components';
+import { PaginationWrapper, PaginationList } from './styles';
+import PerPageChooser from './PerPageChooser';
 
 // pages calculation
 const generatePages = (pageCount, currentPage) => {
@@ -69,43 +64,6 @@ const generatePages = (pageCount, currentPage) => {
   }
 
   return pages;
-};
-
-// per page chooser component
-const PerPageChooser = ({ history }) => {
-  const currentPerPage = Number(router.getParam(history, 'perPage')) || 20;
-
-  const onClick = perPage => {
-    router.setParams(history, { perPage });
-  };
-
-  const renderOption = number => {
-    return (
-      <Option>
-        <a onClick={() => onClick(number)}>{number}</a>
-      </Option>
-    );
-  };
-
-  return (
-    <Dropdown id="per-page-chooser" className="dropup">
-      <DropdownToggle bsRole="toggle">
-        <PerPageButton>
-          {currentPerPage} per page <Icon icon="ios-arrow-up" />
-        </PerPageButton>
-      </DropdownToggle>
-      <Dropdown.Menu>
-        {renderOption(20)}
-        {renderOption(50)}
-        {renderOption(100)}
-        {renderOption(200)}
-      </Dropdown.Menu>
-    </Dropdown>
-  );
-};
-
-PerPageChooser.propTypes = {
-  history: PropTypes.object.isRequired
 };
 
 // page chooser component
